ordering package priority based on drone interests
