+++ .env.example
@@ -15,3 +15,6 @@ BATTERY_DISCHARGING_PICK=1.0
 BATTERY_DISCHARGING_DELIVER=1.0
 BATTERY_CHARGING_RATE=0.3
 BATTERY_EMERGENCY_RATIO=0.5
+
+DISTANCE_PACKAGE_SCORE_WEIGHT=0.2
+DISTANCE_DESTINATION_SCORE_WEIGHT=0.1
