+++ internal/constants/env.go
@@ -26,6 +26,9 @@ var (
 	BATTERY_DISCHARGING_DELIVER = envFloat("BATTERY_DISCHARGING_DELIVER", 1.0)
 	BATTERY_CHARGING_RATE       = envFloat("BATTERY_CHARGING_RATE", 0.3)
 	BATTERY_EMERGENCY_RATIO     = envFloat("BATTERY_EMERGENCY_RATIO", 0.5) // Below x% of battery, will try to recharge
+
+	DISTANCE_PACKAGE_SCORE_WEIGHT     = envFloat("DISTANCE_PACKAGE_SCORE_WEIGHT", 0.2)
+	DISTANCE_DESTINATION_SCORE_WEIGHT     = envFloat("DISTANCE_DESTINATION_SCORE_WEIGHT", 0.1)
 )
 
 func envInt(key string, def int) int {
